+++ app/src/main/java/com/anytypeio/anytype/di/feature/OtherSettingsDI.kt
@@ -2,7 +2,9 @@ package com.anytypeio.anytype.di.feature
 
 import com.anytypeio.anytype.analytics.base.Analytics
 import com.anytypeio.anytype.core_utils.di.scope.PerScreen
+import com.anytypeio.anytype.domain.block.repo.BlockRepository
 import com.anytypeio.anytype.domain.config.UserSettingsRepository
+import com.anytypeio.anytype.domain.device.ClearFileCache
 import com.anytypeio.anytype.domain.launch.GetDefaultEditorType
 import com.anytypeio.anytype.domain.launch.SetDefaultEditorType
 import com.anytypeio.anytype.domain.misc.AppActionManager
@@ -41,16 +43,23 @@ object OtherSettingsModule {
         SetDefaultEditorType(repo)
 
     @JvmStatic
+    @PerScreen
+    @Provides
+    fun provideClearFileCache(repo: BlockRepository): ClearFileCache = ClearFileCache(repo)
+
+    @JvmStatic
     @Provides
     @PerScreen
     fun provideOtherSettingsFactory(
         getDefaultEditorType: GetDefaultEditorType,
         setDefaultEditorType: SetDefaultEditorType,
+        clearFileCache: ClearFileCache,
         appActionManager: AppActionManager,
         analytics: Analytics
     ): OtherSettingsViewModel.Factory = OtherSettingsViewModel.Factory(
         getDefaultEditorType = getDefaultEditorType,
         setDefaultEditorType = setDefaultEditorType,
+        clearFileCache = clearFileCache,
         appActionManager = appActionManager,
         analytics = analytics
     )