+++ presentation/src/main/java/com/anytypeio/anytype/presentation/settings/OtherSettingsViewModel.kt
@@ -10,21 +10,28 @@ import com.anytypeio.anytype.analytics.event.EventAnalytics
 import com.anytypeio.anytype.analytics.props.Props
 import com.anytypeio.anytype.core_models.Id
 import com.anytypeio.anytype.core_models.SmartBlockType
+import com.anytypeio.anytype.domain.base.BaseUseCase
+import com.anytypeio.anytype.domain.base.Interactor
+import com.anytypeio.anytype.domain.device.ClearFileCache
 import com.anytypeio.anytype.domain.launch.GetDefaultEditorType
 import com.anytypeio.anytype.domain.launch.SetDefaultEditorType
 import com.anytypeio.anytype.domain.misc.AppActionManager
 import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import timber.log.Timber
 
 class OtherSettingsViewModel(
     private val getDefaultEditorType: GetDefaultEditorType,
     private val setDefaultEditorType: SetDefaultEditorType,
+    private val clearFileCache: ClearFileCache,
     private val appActionManager: AppActionManager,
     private val analytics: Analytics
 ) : ViewModel() {
 
     val commands = MutableSharedFlow<Command>(replay = 0)
+    val isClearFileCacheInProgress = MutableStateFlow(false)
 
     init {
         viewModelScope.launch {
@@ -38,6 +45,27 @@ class OtherSettingsViewModel(
     fun onObjectTypeClicked() {
         viewModelScope.launch {
             commands.emit(Command.NavigateToObjectTypesScreen(DEFAULT_SETTINGS_SMART_BLOCK_TYPE))
+        }
+    }
+
+    fun onClearFileCacheClicked() {
+        viewModelScope.launch {
+            clearFileCache(BaseUseCase.None).collect { status ->
+                when (status) {
+                    is Interactor.Status.Started -> {
+                        isClearFileCacheInProgress.value = true
+                    }
+                    is Interactor.Status.Error -> {
+                        isClearFileCacheInProgress.value = false
+                        val msg = "Error while clearing file cache: ${status.throwable.message}"
+                        Timber.e(status.throwable, "Error while clearing file cache")
+                        commands.emit(Command.Toast(msg))
+                    }
+                    Interactor.Status.Success -> {
+                        isClearFileCacheInProgress.value = false
+                    }
+                }
+            }
         }
     }
 
@@ -88,6 +116,7 @@ class OtherSettingsViewModel(
     class Factory(
         private val getDefaultEditorType: GetDefaultEditorType,
         private val setDefaultEditorType: SetDefaultEditorType,
+        private val clearFileCache: ClearFileCache,
         private val appActionManager: AppActionManager,
         private val analytics: Analytics
     ) : ViewModelProvider.Factory {
@@ -97,6 +126,7 @@ class OtherSettingsViewModel(
         ): T = OtherSettingsViewModel(
             getDefaultEditorType = getDefaultEditorType,
             setDefaultEditorType = setDefaultEditorType,
+            clearFileCache = clearFileCache,
             appActionManager = appActionManager,
             analytics = analytics
         ) as T